+++ core/domain/src/main/java/com/same/alarm/domain/usecase/alarm/AutoStopAlarmUseCase.kt
@@ -1,4 +1,4 @@
-package com.same.alarm.domain.usecase
+package com.same.alarm.domain.usecase.alarm
 
 import com.same.alarm.common.AlarmConstants.REPEAT_COUNT
 import com.same.alarm.domain.repository.AlarmHelper
