+++ core/domain/src/main/java/com/same/alarm/domain/usecase/alarm/LoadAlarmUseCase.kt
@@ -1,4 +1,4 @@
-package com.same.alarm.domain.usecase
+package com.same.alarm.domain.usecase.alarm
 
 import com.same.alarm.domain.repository.AlarmRepository
 import javax.inject.Inject
