+++ core/domain/src/main/java/com/same/alarm/domain/usecase/alarm/ToggleAlarmUseCase.kt
@@ -1,16 +1,20 @@
-package com.same.alarm.domain.usecase
+package com.same.alarm.domain.usecase.alarm
 
 import com.same.alarm.domain.repository.AlarmHelper
 import com.same.alarm.domain.repository.AlarmRepository
 import com.same.alarm.model.Alarm
 import javax.inject.Inject
 
-class RemoveAlarmUseCase @Inject constructor(
+class ToggleAlarmUseCase @Inject constructor(
     private val alarmRepository: AlarmRepository,
     private val alarmHelper: AlarmHelper
 ) {
     suspend operator fun invoke(alarm: Alarm) {
-        alarmRepository.removeAlarm(alarm)
-        alarmHelper.unScheduleAlarm(alarm)
+        alarmRepository.updateAlarm(alarm)
+        if (alarm.isActive) {
+            alarmHelper.scheduleAlarm(alarm)
+        } else {
+            alarmHelper.unScheduleAlarm(alarm)
+        }
     }
 }
