+++ core/domain/src/main/java/com/same/alarm/domain/usecase/alarm/UpdateAlarmUseCase.kt
@@ -1,20 +1,19 @@
-package com.same.alarm.domain.usecase
+package com.same.alarm.domain.usecase.alarm
 
 import com.same.alarm.domain.repository.AlarmHelper
 import com.same.alarm.domain.repository.AlarmRepository
 import com.same.alarm.model.Alarm
 import javax.inject.Inject
 
-class ToggleAlarmUseCase @Inject constructor(
+class UpdateAlarmUseCase @Inject constructor(
     private val alarmRepository: AlarmRepository,
     private val alarmHelper: AlarmHelper
 ) {
-    suspend operator fun invoke(alarm: Alarm) {
-        alarmRepository.updateAlarm(alarm)
-        if (alarm.isActive) {
-            alarmHelper.scheduleAlarm(alarm)
-        } else {
+    suspend operator fun invoke(alarm: Alarm): Result<Unit> {
+        return runCatching {
             alarmHelper.unScheduleAlarm(alarm)
+            alarmRepository.updateAlarm(alarm)
+            alarmHelper.scheduleAlarm(alarm)
         }
     }
 }
