+++ feature/edit/src/main/java/com/same/alarm/edit/EditViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
-import com.same.alarm.domain.usecase.LoadAlarmUseCase
-import com.same.alarm.domain.usecase.UpdateAlarmUseCase
+import com.same.alarm.domain.usecase.alarm.LoadAlarmUseCase
+import com.same.alarm.domain.usecase.alarm.UpdateAlarmUseCase
 import com.same.alarm.model.Alarm
 import com.same.alarm.navigation.Route
 import dagger.hilt.android.lifecycle.HiltViewModel
